requirements.txt
Flask==2.1.2
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.4.4
python-dotenv==0.19.2
bcrypt==3.2.0
pymongo==3.12.0
